Rename consensus service to carnot consensus (#556)
