+++ nodes/nomos-node/Cargo.toml
@@ -25,7 +25,7 @@ nomos-api = { path = "../../nomos-services/api" }
 nomos-log = { path = "../../nomos-services/log" }
 nomos-mempool = { path = "../../nomos-services/mempool", features = ["mock", "libp2p"] }
 nomos-http = { path = "../../nomos-services/http", features = ["http"] }
-nomos-consensus = { path = "../../nomos-services/consensus", features = ["libp2p"] }
+carnot-consensus = { path = "../../nomos-services/carnot-consensus", features = ["libp2p"] }
 nomos-storage = { path = "../../nomos-services/storage", features = ["sled"] }
 nomos-libp2p = { path = "../../nomos-libp2p" }
 nomos-da = { path = "../../nomos-services/data-availability", features = ["libp2p"] }
