+++ nomos-services/api/Cargo.toml
@@ -15,7 +15,7 @@ tracing = "0.1"
 
 carnot-engine = { path = "../../consensus/carnot-engine" }
 nomos-core = { path = "../../nomos-core" }
-nomos-consensus = { path = "../../nomos-services/consensus" }
+carnot-consensus = { path = "../carnot-consensus" }
 nomos-network = { path = "../../nomos-services/network" }
 nomos-da = { path = "../../nomos-services/data-availability" }
 nomos-mempool = { path = "../../nomos-services/mempool", features = ["mock", "libp2p", "openapi"] }
