+++ nomos-services/carnot-consensus/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "nomos-consensus"
+name = "carnot-consensus"
 version = "0.1.0"
 edition = "2021"
