+++ nomos-services/carnot-consensus/src/committee_membership/mod.rs
