+++ nomos-services/carnot-consensus/src/lib.rs
