+++ nomos-services/carnot-consensus/src/network/adapters/mock.rs
