+++ nomos-services/carnot-consensus/src/network/adapters/mod.rs
