+++ nomos-services/carnot-consensus/src/tally/happy.rs
