+++ nomos-services/carnot-consensus/src/tally/mod.rs
