+++ tests/Cargo.toml
@@ -6,7 +6,7 @@ publish = false
 
 [dependencies]
 nomos-node = { path = "../nodes/nomos-node", default-features = false }
-nomos-consensus = { path = "../nomos-services/consensus" }
+carnot-consensus = { path = "../nomos-services/carnot-consensus" }
 nomos-network = { path = "../nomos-services/network", features = ["libp2p"]}
 nomos-log = { path = "../nomos-services/log" }
 nomos-api = { path = "../nomos-services/api" }
