+++ tests/src/tests/unhappy.rs
@@ -1,7 +1,7 @@
+use carnot_consensus::CarnotInfo;
 use carnot_engine::{Block, NodeId, TimeoutQc, View};
 use fraction::Fraction;
 use futures::stream::{self, StreamExt};
-use nomos_consensus::CarnotInfo;
 use std::{collections::HashSet, time::Duration};
 use tests::{adjust_timeout, ConsensusConfig, MixNode, Node, NomosNode, SpawnConfig};
